Create Project 1 : Password Generator
A password is a combination of lowercase, uppercase, symbols, and
digits. An ideal or a strong password contains a combination of these
four.
 Import random: To randomly generate a password we use this module.
Random is a built-in module used to generate a random subset or a
substring of a list, array or string. In our case, it will generate a random
string containing characters from the original string.
 Get User Input
o Ask the user for the desired length of the password.
o Optionally, ask the user for specific requirements (e.g., include/exclude
special characters, minimum number of digits, etc.).
o Ensure the user inputs a valid number for the password length.
o Allow users to specify minimum counts for each character type (e.g., at least 2
digits).