Create PROJECT 1: SUDOKU GAME
The Sudoku puzzle game is played on a 9 x 9 grid. Within the
rows and columns are 9 boxes which are made up of 3 x 3 space.
The purpose of the puzzle is to enter a digit number from 1 to 9
in each row & a 9 × 9 grid column without duplicating any
numbers between the row, column or box.